% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Compute ell-value, u-value, and h-value}
\usage{
accuracy(result, zeta = seq(from = 0, to = 0.8, by = 0.05))
}
\arguments{
\item{result}{an object of class “MLcovar”, usually, a result of a call to \code{\link{MLcovar}}.}

\item{zeta}{The vector of variance reduction factor (\link{0, 1}). Default is seq(from = 0, to = 0.8, by = 0.05).}
}
\description{
Compute ell-value, u-value, and h-value
}
