% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MLcovar-package.R
\docType{package}
\name{MLCovar-package}
\alias{MLCovar}
\alias{MLCovar-package}
\title{Package documentation}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
