% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{MLcovar}
\alias{MLcovar}
\title{Proposed estimator}
\usage{
MLcovar(main_model, proxy_model, data, labeled_set_var_name, n_boot = 100)
}
\arguments{
\item{main_model}{A function that estimates the target parameter with the
labeled data. The function takes the data as the only argument and returns
the estimate of the target parameter. The target parameter must be a scalar.}

\item{proxy_model}{A function that estimates the target parameter with the
proxy variables. The function takes the data as the only argument and
returns the estimate of the target parameter. The target parameter must be a
scalar or a vector.}

\item{data}{A data frame that contains the labeled and unlabeled data. The
rows that contains the labeled data must be indicated by a variable in the
data frame. The variable name must be provided in the labeled_set_var_name
argument.}

\item{labeled_set_var_name}{The name of the variable that indicates whether
rows are labeled or not. The variable must be binary.}
}
\description{
Proposed estimator
}
