% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peri-package.R
\docType{package}
\name{peri-package}
\alias{peri-package}
\title{Package documentation}
\description{
Provides a framework for prediction-error robust inference for statistical analysis. The package implements the method proposed in Katsumata and Yamauchi (2025).
}
\author{
\strong{Maintainer}: Soichiro Yamauchi \email{soichiro@alumni.princeton.edu} (\href{https://orcid.org/0000-0003-0554-2717}{ORCID})

Authors:
\itemize{
  \item Hiroto Katsumata \email{hrt.katsumata@gmail.com} (\href{https://orcid.org/0000-0001-5901-8844}{ORCID})
}

}
\keyword{internal}
