% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.accuracy.R
\name{plot.accuracy}
\alias{plot.accuracy}
\title{Plot results of the accuracy analysis}
\usage{
\method{plot}{accuracy}(x, ...)
}
\arguments{
\item{x}{an object of class “accuracy”, usually, a result of a call to \code{\link{accuracy}}.}

\item{...}{other graphical parameters (see \code{\link[graphics]{par}}).}
}
\description{
Plot results of the accuracy analysis
}
