% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.MLcovar}
\alias{print.summary.MLcovar}
\title{Print summary for MLcovar results}
\usage{
\method{print}{summary.MLcovar}(x, ...)
}
\arguments{
\item{x}{An object of class "summary.MLcovar" returned by the summary.MLcovar
function.}

\item{...}{Additional arguments.}
}
\description{
Prints the summary of the results from the MLcovar function.
}
