% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.peri}
\alias{print.summary.peri}
\title{Print summary for peri results}
\usage{
\method{print}{summary.peri}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.peri" returned by the summary.peri
function.}

\item{...}{Additional arguments.}
}
\description{
Prints the summary of the results from the peri function.
}
