% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.prisa}
\alias{print.summary.prisa}
\title{Print summary for prisa results}
\usage{
\method{print}{summary.prisa}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class "summary.prisa" returned by the summary.prisa
function.}

\item{digits}{The number of significant digits to print. Default is 4.}

\item{...}{Additional arguments.}
}
\description{
Prints the summary of the results from the prisa function.
}
