% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prisa-lm.R
\name{prisa_lm}
\alias{prisa_lm}
\title{Implement the Prediction-error Robust Inference for linear regression}
\usage{
prisa_lm(
  main_formula,
  proxy_formula,
  data,
  labeled_set_var_name,
  options = SetOptions(n_boot = 500, use_full = TRUE, use_parallel = TRUE, n_cores =
    parallel::detectCores() - 1, cluster_var_name = NULL),
  args_main_model = list(),
  args_proxy_model = list()
)
}
\description{
Implement the Prediction-error Robust Inference for linear regression
}
