% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SetOptions}
\alias{SetOptions}
\title{Set options for the main function}
\usage{
SetOptions(
  n_boot = 500,
  use_full = TRUE,
  is_parallel = TRUE,
  seed_value = 1234
)
}
\arguments{
\item{n_boot}{The number of bootstrap samples for the labeled data. Default
is 500. When use_full is TRUE, the full data is also used for the bootstrap.}

\item{use_full}{A logical value that indicates whether the full data should
be used in the proxy model. Default is TRUE. If FALSE, the unlabeled data
will be used in the proxy model.}

\item{is_parallel}{A logical value that indicates whether the bootstrap
should be run in parallel. Default is TRUE.}

\item{seed_value}{The seed value for the random number generator. Default is
1234.}
}
\value{
A named list of options.
}
\description{
Returns a list of configuration options to be passed to \code{\link[=MLcovar]{MLcovar()}}.
}
\seealso{
\code{\link[=MLcovar]{MLcovar()}}
}
