% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{summary.accuracy}
\alias{summary.accuracy}
\title{Summary for accuracy analysis}
\usage{
\method{summary}{accuracy}(object, digits = 4, nrows = 10, ...)
}
\arguments{
\item{object}{An object of class "accuracy" returned by accuracy().}

\item{digits}{Number of significant digits to print. Default is 4.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary for accuracy analysis
}
