% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.peri}
\alias{summary.peri}
\title{Summary for MLcovar results
#' @description Provides a summary of the results from the MLcovar function.}
\usage{
\method{summary}{peri}(object, ...)
}
\arguments{
\item{object}{An object of class "MLcovar" returned by the MLcovar function.}

\item{...}{Additional arguments.}
}
\value{
A list containing the summary of the results.
}
\description{
Summary for MLcovar results
#' @description Provides a summary of the results from the MLcovar function.
}
