% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.prisa}
\alias{summary.prisa}
\title{Summary for prisa results}
\usage{
\method{summary}{prisa}(object, ...)
}
\arguments{
\item{object}{An object of class "prisa" returned by the prisa function.}

\item{...}{Additional arguments.}
}
\value{
A list containing the summary of the results.
}
\description{
Provides a summary of the results from the prisa function.
}
