% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Compute ell-value, u-value, and h-value}
\usage{
accuracy(result, zeta = seq(from = 0.05, to = 0.8, by = 0.05))
}
\arguments{
\item{result}{an object of class “MLcovar”, usually, a result of a call to \code{\link{MLcovar}}.}

\item{zeta}{A vector of variance reduction factor (\link{0, 1}).
Default is between 0.05 and 0.8.}
}
\description{
Compute ell-value, u-value, and h-value
}
\examples{
set.seed(1234)
n_ell <- 200
n_u <- 5000
n_all <- n_ell + n_u

# true model
X <- rnorm(n_all)
Y <- 0.5 + X + rnorm(n_all, mean = 0, sd = 0.3)

# Add noise to the outcome to create proxy variable
# Y_proxy are Y with non-classical measurement errors
Y_proxy_1 <- Y + rnorm(n_all, mean = X, sd = 0.1)

# Create a data frame
df_test <- data.frame(
  Y = Y,
  Y_proxy_1 = Y_proxy_1,
  X = X,
  is_labeled = c(rep(1, n_ell), rep(0, n_u))
) |> 
  dplyr::mutate(Y = dplyr::if_else(is_labeled == 1, Y, NA_real_))

# Prepare functions for the analysis
fn_main <- function(df) {
  fit <- lm(Y ~ X, data = df)
  return(fit$coef[2])
}

fn_proxy_1 <- function(df) {
  fit <- lm(Y_proxy_1 ~ X, data = df)
  return(fit$coef)
}

# Run functions
fit_1 <- MLcovar(
  main_model = fn_main,
  proxy_model = fn_proxy_1,
  data = df_test,
  labeled_set_var_name = "is_labeled",
  n_boot = 5000,
  use_full = TRUE
)
summary(fit_1)

accuracy_out <- accuracy(fit_1)
plot(accuracy_out)
}
