% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{SetOptions}
\alias{SetOptions}
\title{Set options for the main function}
\usage{
SetOptions(
  n_boot = 500,
  use_full = TRUE,
  use_parallel = TRUE,
  n_cores = parallel::detectCores() - 1,
  seed_value = floor(runif(1, 1, 1e+07))
)
}
\arguments{
\item{n_boot}{The number of bootstrap samples for the labeled data. Default
is 500. When use_full is TRUE, the full data is also used for the bootstrap.}

\item{use_full}{A logical value that indicates whether the full data should
be used in the proxy model. Default is TRUE. If FALSE, the unlabeled data
will be used in the proxy model.}

\item{use_parallel}{A logical value that indicates whether the bootstrap
should be run in parallel. Default is TRUE.}

\item{n_cores}{The number of cores to be used for bootstrap interactions.
Default is the number of cores available minus one. This value will be
ignored if use_parallel is FALSE.}

\item{seed_value}{The seed value for the random number generator. Default is
drawn from a uniform between 1 and 1e7.}
}
\value{
A named list of options.
}
\description{
Returns a list of configuration options to be passed to \code{\link[=MLcovar]{MLcovar()}}.
}
\seealso{
\code{\link[=MLcovar]{MLcovar()}}
}
